Extract the script from MakersEditView.vue into a composable
